factor out query_partitions from merge
